try/catch binary object deletion, delete only if not empty

+++ conductor-client/src/main/kotlin/com/openlattice/data/jobs/DataDeletionJob.kt
@@ -45,7 +45,6 @@ import com.zaxxer.hikari.HikariDataSource
 import org.apache.olingo.commons.api.edm.EdmPrimitiveTypeKind
 import java.sql.PreparedStatement
 import java.util.*
-import java.util.concurrent.atomic.AtomicLong
 
 class DataDeletionJob(
         state: DataDeletionJobState
@@ -293,7 +292,6 @@ class DataDeletionJob(
             entityKeyIds: Collection<UUID>,
             propertyTypeIds: Collection<UUID>
     ) {
-        val count = AtomicLong()
         val s3Keys = BasePostgresIterable<String>(
                 PreparedStatementHolderSupplier(hds, selectEntitiesTextProperties, FETCH_SIZE) { ps ->
                     val connection = ps.connection
@@ -306,8 +304,13 @@ class DataDeletionJob(
                 }
         ) { it.getString(getMergedDataColumnName(PostgresDatatype.TEXT)) }.toList()
 
-        byteBlobDataManager.deleteObjects(s3Keys)
-        count.addAndGet(s3Keys.size.toLong())
+        if (s3Keys.isNotEmpty()) {
+            try {
+                byteBlobDataManager.deleteObjects(s3Keys)
+            } catch (e: Exception) {
+                logger.error("Unable to delete object from s3 for entity sets {} with ids {}", entitySetIds, entityKeyIds, e)
+            }
+        }
     }
 
     @JsonIgnore
